# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Medaware Catalyst
  description: The API Specification for Medaware
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development environment
paths:
...
  /public/article:
    get:
      operationId: catalystGetArticle
      summary: Retrieve a full article by ID
      tags:
        - Public
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - title
                  - lead
                  - html
                properties:
                  title:
                    type: string
                  lead:
                    type: string
                  html:
                    type: string
        404:
          description: Article not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/catalystError"
  /public/search:
    get:
      operationId: catalystQueryArticle
      summary: Search existing articles by querying lucene
      tags:
        - Public
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/articleSearchResponse"
components:
  schemas:
    catalystError:
      type: object
      required:
        - brief
        - message
        - status
      properties:
        brief:
          type: string
        message:
          type: string
        status:
          type: integer
...
    briefArticleResponse:
      type: object
      required:
        - title
        - lead
        - id
      properties:
        title:
          type: string
        lead:
          type: string
        id:
          type: string
          format: uuid
    articleSearchResponse:
      type: array
      items:
        $ref: "#/components/schemas/briefArticleResponse"